# docs/generated/openapi.json
{
  "components": {
    "schemas": {
...
      "DifficultyLevel": {
        "enum": [
          "EASY",
          "MEDIUM",
          "HARD"
        ],
        "title": "DifficultyLevel",
        "type": "string"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "OptionBase": {
        "properties": {
          "content": {
            "title": "Content",
            "type": "string"
          },
          "is_correct": {
            "title": "Is Correct",
            "type": "boolean"
          }
        },
        "required": [
          "content",
          "is_correct"
        ],
        "title": "OptionBase",
        "type": "object"
      },
      "OptionResponse": {
        "properties": {
          "content": {
            "title": "Content",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "is_correct": {
            "title": "Is Correct",
            "type": "boolean"
          }
        },
        "required": [
          "id",
          "content",
          "is_correct"
        ],
        "title": "OptionResponse",
        "type": "object"
      },
...
      "Page_QuestionResponse_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/QuestionResponse"
            },
            "title": "Items",
            "type": "array"
          },
          "page": {
            "minimum": 1,
            "title": "Page",
            "type": "integer"
          },
          "pages": {
            "minimum": 0,
            "title": "Pages",
            "type": "integer"
          },
          "size": {
            "minimum": 1,
            "title": "Size",
            "type": "integer"
          },
          "total": {
            "minimum": 0,
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "total",
          "page",
          "size",
          "pages"
        ],
        "title": "Page[QuestionResponse]",
        "type": "object"
      },
...
      "QuestionBase": {
        "properties": {
          "content": {
            "title": "Content",
            "type": "string"
          },
          "difficulty": {
            "$ref": "#/components/schemas/DifficultyLevel",
            "default": "MEDIUM"
          },
          "metadata_json": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Metadata Json"
          },
          "options": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/OptionBase"
            },
            "title": "Options",
            "type": "array"
          },
          "points": {
            "default": 1,
            "title": "Points",
            "type": "number"
          },
          "question_type": {
            "$ref": "#/components/schemas/QuestionType",
            "default": "MULTIPLE_CHOICE"
          },
          "topic_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Topic Id"
          }
        },
        "required": [
          "content"
        ],
        "title": "QuestionBase",
        "type": "object"
      },
...
      "QuestionResponse": {
        "properties": {
          "content": {
            "title": "Content",
            "type": "string"
          },
          "difficulty": {
            "$ref": "#/components/schemas/DifficultyLevel"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "is_ai_generated": {
            "title": "Is Ai Generated",
            "type": "boolean"
          },
          "metadata_json": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Metadata Json"
          },
          "options": {
            "items": {
              "$ref": "#/components/schemas/OptionResponse"
            },
            "title": "Options",
            "type": "array"
          },
          "points": {
            "title": "Points",
            "type": "number"
          },
          "question_type": {
            "$ref": "#/components/schemas/QuestionType"
          },
          "topic_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Topic Id"
          }
        },
        "required": [
          "id",
          "content",
          "is_ai_generated",
          "points",
          "question_type",
          "difficulty",
          "options"
        ],
        "title": "QuestionResponse",
        "type": "object"
      },
...
      "QuestionType": {
        "enum": [
          "SINGLE_CHOICE",
          "MULTIPLE_CHOICE",
          "MATCHING",
          "FILL_IN_BLANK"
        ],
        "title": "QuestionType",
        "type": "string"
      },
      "QuestionUpdate": {
        "properties": {
          "content": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Content"
          },
          "difficulty": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/DifficultyLevel"
              },
              {
                "type": "null"
              }
            ]
          },
          "metadata_json": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Metadata Json"
          },
          "options": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/OptionBase"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Options"
          },
          "points": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Points"
          },
          "question_type": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/QuestionType"
              },
              {
                "type": "null"
              }
            ]
          },
          "topic_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Topic Id"
          }
        },
        "title": "QuestionUpdate",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/auth/login"
          }
        },
        "type": "oauth2"
      }
    }
  },
  "info": {
    "title": "AI Quiz System",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/exams/{exam_id}/questions": {
      "post": {
        "operationId": "create_question_in_exam_exams__exam_id__questions_post",
        "parameters": [
          {
            "in": "path",
            "name": "exam_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Exam Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QuestionBase"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QuestionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Create Question In Exam",
        "tags": [
          "exams"
        ]
      }
    },
...
    "/materials/{material_id}/questions": {
      "get": {
        "operationId": "get_ai_questions_materials__material_id__questions_get",
        "parameters": [
          {
            "in": "path",
            "name": "material_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Material Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/QuestionResponse"
                  },
                  "title": "Response Get Ai Questions Materials  Material Id  Questions Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Get Ai Questions",
        "tags": [
          "materials"
        ]
      }
    },
...
    "/questions": {
      "get": {
        "description": "Get paginated question bank list with eager-loaded options and filters.",
        "operationId": "read_questions_questions_get",
        "parameters": [
          {
            "in": "query",
            "name": "topic_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "uuid",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Topic Id"
            }
          },
          {
            "in": "query",
            "name": "difficulty",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Difficulty"
            }
          },
          {
            "in": "query",
            "name": "search",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Search"
            }
          },
          {
            "description": "Page number",
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "description": "Page number",
              "minimum": 1,
              "title": "Page",
              "type": "integer"
            }
          },
          {
            "description": "Page size",
            "in": "query",
            "name": "size",
            "required": false,
            "schema": {
              "default": 50,
              "description": "Page size",
              "maximum": 100,
              "minimum": 1,
              "title": "Size",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_QuestionResponse_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Read Questions",
        "tags": [
          "questions"
        ]
      },
      "post": {
        "description": "Create a standalone question bank question with options.",
        "operationId": "create_question_questions_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QuestionBase"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QuestionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Create Question",
        "tags": [
          "questions"
        ]
      }
    },
    "/questions/{question_id}": {
      "delete": {
        "operationId": "delete_question_questions__question_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "question_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Question Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Delete Question",
        "tags": [
          "questions"
        ]
      },
      "get": {
        "operationId": "get_question_questions__question_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "question_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Question Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QuestionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Get Question",
        "tags": [
          "questions"
        ]
      },
      "put": {
        "operationId": "update_question_questions__question_id__put",
        "parameters": [
          {
            "in": "path",
            "name": "question_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Question Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QuestionUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QuestionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Update Question",
        "tags": [
          "questions"
        ]
      }
    },
    "/questions/{question_id}/restore": {
      "post": {
        "description": "Restore a soft-deleted question within the 30-day retention window.\n(Requires owner or admin authentication)",
        "operationId": "restore_question_questions__question_id__restore_post",
        "parameters": [
          {
            "in": "path",
            "name": "question_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Question Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QuestionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Restore Question",
        "tags": [
          "questions"
        ]
      }
    },
...
  }
}